Fixed problem with updating a filter after time outs.
After the initial "upload" of the first query results
the Finder might keep these results in memory even after
we cleared this filter with "stopSearch:".
So the Finder might ask for attributes and so on of results
from the last query. It is just a short time and the Finder
will eventually notice that these files are not around
any more, but we have to deal with this anyway. Otherwise
the Finder will show dead files for these and you will
get a "spinning wheel" if you want to access file properties.

So to solve this situation we will keep track of the last
dropped path (such that we can find these old results on
the hard drive) and just give back attributes of files from
the last result set if the Finder asks for it.
Therfore this method will give back a "checkedRealPath"
even if we do not longer have a pathContainer.